agent/requirements.txt
beautifulsoup4
boto3
llama-index
openai
python-dotenv
requests
tenacity
termcolor 